Separated up author metadata entry and changed display.
The initial concept of using one panel for everything did not work
well with the pop-up keyboard of mobile devices. So there are now
two panels, for thumbnail and links. The notes need to be added.

Implemented the first layout for displying this data, but it need to be
changed for medium size devices like iPads.